requirements.txt
python-telegram-bot==20.3
psycopg2-binary==2.9.9
reportlab==4.0.9
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
